+++ tests/test_connections.py
@@ -72,20 +72,65 @@ class TestConnections:
                         reason="requires libvips >= 8.9")
     def test_target_custom(self):
         filename = temp_filename(self.tempdir, '.png')
-        output_file = open(filename, "wb")
+        output_file = open(filename, "w+b")
 
         def write_handler(chunk):
             return output_file.write(chunk)
 
-        def finish_handler():
-            output_file.close()
+        def end_handler():
+            try:
+                output_file.close()
+            except IOError:
+                return -1
+            else:
+                return 0
 
         target = pyvips.TargetCustom()
         target.on_write(write_handler)
-        target.on_finish(finish_handler)
+        target.on_end(end_handler)
 
         image = pyvips.Image.new_from_file(JPEG_FILE, access='sequential')
         image.write_to_target(target, '.png')
+
+        image = pyvips.Image.new_from_file(JPEG_FILE, access='sequential')
+        image2 = pyvips.Image.new_from_file(filename, access='sequential')
+
+        assert (image - image2).abs().max() == 0
+
+    @skip_if_no('jpegload')
+    @skip_if_no('tiffsave')
+    @pytest.mark.skipif(not pyvips.at_least_libvips(8, 13),
+                        reason="requires libvips >= 8.13")
+    def test_target_custom_seek(self):
+        filename = temp_filename(self.tempdir, '.png')
+        output_file = open(filename, "w+b")
+
+        def write_handler(chunk):
+            return output_file.write(chunk)
+
+        def read_handler(size):
+            return output_file.read(size)
+
+        def seek_handler(offset, whence):
+            output_file.seek(offset, whence)
+            return output_file.tell()
+
+        def end_handler():
+            try:
+                output_file.close()
+            except IOError:
+                return -1
+            else:
+                return 0
+
+        target = pyvips.TargetCustom()
+        target.on_write(write_handler)
+        target.on_read(read_handler)
+        target.on_seek(seek_handler)
+        target.on_end(end_handler)
+
+        image = pyvips.Image.new_from_file(JPEG_FILE, access='sequential')
+        image.write_to_target(target, '.tif')
 
         image = pyvips.Image.new_from_file(JPEG_FILE, access='sequential')
         image2 = pyvips.Image.new_from_file(filename, access='sequential')